Remove unnecessary code
`existingEntities` is populated in init stage in `#getEntityNames` so no
need to do it again.

`auditedEntities` is not used anymore as it was just a copy of
`entitiesToUpdate`.

The `JaversEntityAuditResource.java` template is already rendered in
writing stage in `#writeAuditPageFiles` so no need to do it again.

Also this code was never actually called as the `#updateEntityAudit`
function was always called with `updateIndex === undefined`.