Update README.md
The repo had been moved from `jhipster` to `hipster-labs` but there was a reference to the old one in the sample code to use and unreleased version.
This is minor but I had to document it on my end for other team members.  Might as well update it for everyone at the source.